create data loading utils script and move data loading from generic into new utils
